change mime type to metadata map for DataPayload + add "help" to Param
DataPayload benefits from being able to provide additional information
about the data blob being sent.  Just supporting mime type was limiting
when we could provide metadata map that might indicate new filenames for
multi-file responses, mime type, file name, total file size, and any
other information that might be relevant for the Client to make use of.

Help for Param is a place to provide more information about the
parameter (or command) to describe its purpose in more detail to an
end-user. This can be shown by the client either as a tool tip (if
brief) or within a help dialog if more complex.