Update full_service.cpp implementation
Update full_service.cpp implementation to make it more friendly to the end user.

This is a proposal/draft written without tests - feel free to make changes and help to develop BKM for code sharing.

It would be the best option to share some common entry points to this SDK API so that it can be included without fear of breaking changes.

Great work by the way!