Several updates here
Support for external object requests in the example device - checks for
files whose name match the requested OID in the working directory first
then falls back to the internal package.

Implementation of multi-language support (and changes to the menu group
builder to allow access to the Protobuf Builder objects that get created
so additional language strings can be added).

Change location of internal files so that we can put command assets and
external object assets in the same package.

Update OAuth interceptor to provide a more appropriate error status