fixed some issues with minimal set and susbscriptions
Should correctly return the minimal set even if for some reason subscriptions are "off" - note that audio deck does NOT have them on. This is done on purpose. But it was worth checking for this edge case.

So if subscriptions ARE on, detail level is subscriptions, and there are no subscriptions, it should only send the device itself - subscriptions can still be added even if they are "off" in the device model.